Rename polymer-project.js to project.js
