+++ generators/app/gulp-tasks/project.js
